+++ src/data_id.c
@@ -506,7 +506,8 @@ char * data_id_convert_company_prefix(char * id_human_readable, char * id_value)
     if ((int)strlen(id_value) < 2) return NULL;
     char * company_prefix = get_gs1_company_prefix(id_value);
     if (company_prefix == NULL) return NULL;
-    char * company_prefix_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+    char * company_prefix_value =
+		(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
     company_prefix_value[0] = 0;
     decode_strcat(company_prefix_value, "GS1 Company Prefix: ");
     decode_strcat(company_prefix_value, company_prefix);
@@ -524,7 +525,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 {
     if (strstr(id_human_readable, "YYYYMMDDHHMM") != NULL) {
         if ((int)strlen(id_value) >= 12) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* year */
             decode_strcat_char(date_value, id_value[0]);
@@ -559,7 +561,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYYYMMDDYYYYMMDD") != NULL) {
         if ((int)strlen(id_value) >= 16) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* year */
             decode_strcat_char(date_value, id_value[0]);
@@ -600,7 +603,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYYYMMDD") != NULL) {
         if ((int)strlen(id_value) >= 8) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* day */
             decode_strcat_char(date_value, id_value[6]);
@@ -627,7 +631,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYYYDDMM") != NULL) {
         if ((int)strlen(id_value) >= 8) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* day */
             decode_strcat_char(date_value, id_value[4]);
@@ -654,7 +659,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYYYWW") != NULL) {
         if ((int)strlen(id_value) >= 6) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* year */
             decode_strcat_char(date_value, id_value[0]);
@@ -671,7 +677,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYWWYYWW") != NULL) {
         if ((int)strlen(id_value) >= 8) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* year */
             decode_strcat(date_value, "20");
@@ -696,7 +703,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYMMDDHH") != NULL) {
         if ((int)strlen(id_value) >= 8) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* day */
             decode_strcat_char(date_value, id_value[4]);
@@ -727,7 +735,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYMMDD") != NULL) {
         if ((int)strlen(id_value) >= 6) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* day */
             decode_strcat_char(date_value, id_value[4]);
@@ -753,7 +762,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYDDMM") != NULL) {
         if ((int)strlen(id_value) >= 6) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* day */
             decode_strcat_char(date_value, id_value[2]);
@@ -779,7 +789,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "DDMMYY") != NULL) {
         if ((int)strlen(id_value) >= 6) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* day */
             decode_strcat_char(date_value, id_value[0]);
@@ -805,7 +816,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "MMDDYY") != NULL) {
         if ((int)strlen(id_value) >= 6) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* day */
             decode_strcat_char(date_value, id_value[2]);
@@ -832,7 +844,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
     if ((strstr(id_human_readable, "YYJJJHH") != NULL) ||
             (strstr(id_human_readable, "YYDDDHH") != NULL)) {
         if ((int)strlen(id_value) >= 7) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* year */
             decode_strcat(date_value, "20");
@@ -855,7 +868,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
     if ((strstr(id_human_readable, "YYJJJ") != NULL) ||
             (strstr(id_human_readable, "YYDDD") != NULL)) {
         if ((int)strlen(id_value) >= 5) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* year */
             decode_strcat(date_value, "20");
@@ -872,7 +886,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "MMYY") != NULL) {
         if ((int)strlen(id_value) >= 4) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* month */
             char * mnth = month_number_to_text(&id_value[0]);
@@ -894,7 +909,8 @@ char * data_id_convert_date(char * id_human_readable, char * id_value)
 
     if (strstr(id_human_readable, "YYWW") != NULL) {
         if ((int)strlen(id_value) >= 4) {
-            char * date_value = (char*)safemalloc(MAX_DECODE_LENGTH*sizeof(char));
+            char * date_value =
+				(char*)safemalloc((size_t)MAX_DECODE_LENGTH * sizeof(char));
             date_value[0] = 0;
             /* year */
             decode_strcat(date_value, "20");
